polizador/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-ajax-datatable==4.4.5
django-appconf==1.0.6
django-debug-toolbar==4.3.0
django-easy-audit==1.3.6
django-import-export==3.3.7
django-select2==8.1.2
django-storages==1.14.2
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
